requirements.txt
certifi==2020.12.5
click==8.0.1
Flask==1.0.3
importlib-metadata==4.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.20.3
Pillow==8.2.0
six==1.16.0
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1
